Move current app into module
