+++ packages/app/src/common/components/LockedGSol.tsx
@@ -6,7 +6,7 @@ import {
   ArrowPathIcon,
   CheckIcon,
 } from "@heroicons/react/24/solid";
-import { solToCarbon, toFixedWithPrecision, toSol } from "../lib/util";
+import { solToCarbon, toFixedWithPrecision, toSol } from "../utils";
 import dayjs from "dayjs";
 import relativeTime from "dayjs/plugin/relativeTime";
 import Spinner from "./Spinner";
