+++ packages/app/src/common/components/Panel.tsx
