+++ packages/app/src/common/content/tooltips.tsx
