+++ packages/app/src/common/sunriseClientWrapper.ts
@@ -10,7 +10,7 @@ import { type Connection, type PublicKey, Transaction } from "@solana/web3.js";
 import { AnchorProvider } from "@project-serum/anchor";
 import type BN from "bn.js";
 import { type AnchorWallet } from "@solana/wallet-adapter-react";
-import { debounce } from "./util";
+import { debounce } from "./utils";
 import { WalletAdapterNetwork } from "@solana/wallet-adapter-base";
 
 const stage =
