+++ packages/app/src/utils/notifications.tsx
@@ -5,7 +5,7 @@ import {
   XMarkIcon,
 } from "@heroicons/react/24/solid";
 import toast from "react-hot-toast";
-import TweetButton from "../components/TweetButton";
+import TweetButton from "../common/components/TweetButton";
 
 export enum NotificationType {
   success = "success",
