+++ packages/client/src/util.ts
@@ -5,6 +5,7 @@ import {
   PublicKey,
   type TokenAmount,
   type Transaction,
+  LAMPORTS_PER_SOL,
 } from "@solana/web3.js";
 import * as anchor from "@project-serum/anchor";
 import { AnchorProvider, BN } from "@project-serum/anchor";
@@ -18,6 +19,12 @@ import { type Details } from "./types/Details";
 
 // zero bn number
 export const ZERO = new BN(0);
+
+const MAX_NUM_PRECISION = 5;
+
+export const toSol = (lamports: BN, precision = MAX_NUM_PRECISION): number =>
+  lamports.div(new BN(10).pow(new BN(precision))).toNumber() /
+  (LAMPORTS_PER_SOL / 10 ** precision);
 
 export const enum ProgramDerivedAddressSeed {
   G_SOL_MINT_AUTHORITY = "gsol_mint_authority",
